fix: remove newgrp command
'newgrp' command didn't work as expected. It opens a new shell session
where the new group is activated. So in the actual session the group is
not applied unless re-logged-in (is that even a word?). Removing newgrp
command and adding message informing user to manually log out and in
once for the group to become active. More annoying but safer in terms of
stability.