feat(install): replace custom installer with default docker installer
Instead of the manually implemented install script, we now rely internally on the
official Docker install script. The source code for this script is available here [1].
The decision was made because the installer needs regular maintenance (new OS versions
etc) and I just don't have the time for that.

The only minor disadvantage is that the one-time installation of Docker is not quite
as clear and beautiful but the advantage is that, we use the official script, which is
maintained and kept up-to-date by Docker themselves.

[1] https://get.docker.com/