fix(ui): menu and/or header printed twice in powershell
Removed the -r flag from all 'read' calls which seems like the only
plausible cause for this bug. The -r flag was added shortly before the
bug occured for the first time. Additionally updated the internal
functionality of the menu. The menu now runs in a loop and doesn't exit
on invalid input but rather prints a warning message and reprompts for
user input.

Fixes #7