fix(funcs): fix functionality of remove-containers and remove-images
Remove-containers and remove-images used to try to delete only postres
containers and images. It appears that the chosen solutions didn't fully
bring the expected results. Thus, the functions are broadend by this fix
and now don't filter for postgres anymore. Now they simply remove
stopped containers and dangling images. The names in the menu have
updated because of those changes aswell.