fix(container): listing ids that are not tagged as latest
Use a broader approach for listing containers that have postgres as
ancestor image. The new approch relies on grep and awk. The docker
filter 'ancestor', that was being used until now, doesn't show
containers that have a specific tag, so for example, container name
'postgres' would be filtered successfully but 'postgres:13' would be
ignored.