fix(ports): inspect fails when no containers are running
Before loading the reserved ports through `docker inspect` we now check
if there are any containers running in the first place. In case that no
containers were running, the inspect command returned an error because
it received no container ids to inspect.
Now we skip the inspect command if there no containers exist in the first
place.