feat(repo): add option to pick docker hub repo
Added the '-o' / '--official-repo' flag. This flags lets the user toggle
between pulling postgres images either from the official postgres docker
repo or from a private repo. The private repo at the point of writing
this is: nikoksr/postgres. The images are almost identical to the
official images but switch the locals to be german by default. If this
behavior is unwanted, the user can pass the '-o' flag. This does not
change anything inside the app from a users perspective. In the
foreground both repos supply the same options to the user, just in
behind the app behavior is of course altered.

This change also cleans up some parts, introduces a '-h' / '--help' and
'-v' / '--version' flag. Lastly, this bumps the version up to 0.30.0.