fix(docker): infinite wait for container to be up
An earlier version introduced support for custom docker repositories.
Recently it was discovered that the images under nikoksr/postgres caused
the script to wait infinitely for containers to start.

After manual investigation it became obvious however that the containers
and postgres were indeed up and running. The problem was triggered by
the altered encoding that the images in nikoksr/postgres use. They
switch from EN to DE, causing postgres toolchain to also behave
differently.

Until recently we checked the return value of pg_ready for a specific
ENGLISH string. Obviously this does not work well in a german system..

For the fix we adapted the test we use for postres' readyness. Instead
of relying on pg_ready, we now send the simplest select possible into
the database and wait for it to become successful.

Also reduced the interval in which we check for readyness from 1 second
to 50 milliseconds.